Use ssize_t instead of short for line lengths
sthen hit a binary patch containing a 'line' of length > 32kB. This made
the short used for storing the line length wrap and resulted in a buffer
underflow and segfault.  This uses a larger type, which doesn't actually
fix the problem, but makes it much less likely to be hit.

ok florian otto sthen

+++ usr.bin/patch/pch.h
@@ -1,4 +1,4 @@
-/*	$OpenBSD: pch.h,v 1.12 2019/12/09 12:08:42 jca Exp $	*/
+/*	$OpenBSD: pch.h,v 1.13 2019/12/11 20:10:17 jca Exp $	*/
 
 /*
  * patch - a program to apply diffs to original files
@@ -53,7 +53,7 @@ LINENUM		pch_newfirst(void);
 LINENUM		pch_repl_lines(void);
 LINENUM		pch_end(void);
 LINENUM		pch_context(void);
-short		pch_line_len(LINENUM);
+ssize_t		pch_line_len(LINENUM);
 char		pch_char(LINENUM);
 char		*pfetch(LINENUM);
 LINENUM		pch_hunk_beg(void);
